px2dtLDA の更新に対応
